# services/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOME/IP Simulation Studio",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Projects",
        "operationId": "getProjects",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCollection"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOverview"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/authoring": {
      "get": {
        "tags": [
          "authoring"
        ],
        "summary": "Project Authoring",
        "operationId": "getProjectAuthoring",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthoringWorkspace"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/documents/{document_id}": {
      "get": {
        "tags": [
          "authoring"
        ],
        "summary": "Get Document",
        "operationId": "getProjectDocument",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentFile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "authoring"
        ],
        "summary": "Save Document",
        "operationId": "saveProjectDocument",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentWriteResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/authoring/validate": {
      "post": {
        "tags": [
          "authoring"
        ],
        "summary": "Validate Authoring",
        "operationId": "validateProjectAuthoring",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ApiError"
      },
...
      "AuthoringWorkspace": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentRef"
            },
            "type": "array",
            "title": "Documents"
          },
          "validation_status": {
            "type": "string",
            "title": "Validation Status"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "interface_index": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterfaceIndex"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "validation_status"
        ],
        "title": "AuthoringWorkspace"
      },
      "CaptureFile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "wireshark_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wireshark Path"
          }
        },
        "type": "object",
        "required": [
          "name",
          "bytes"
        ],
        "title": "CaptureFile"
      },
      "CaptureSession": {
        "properties": {
          "wireshark_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wireshark Url"
          },
          "wireshark_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wireshark Root"
          },
          "display_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Filter"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/CaptureFile"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "title": "CaptureSession"
      },
      "ContainerState": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "ContainerState"
      },
      "DeploymentSummary": {
        "properties": {
          "document": {
            "type": "string",
            "title": "Document"
          },
          "ids": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Ids"
          },
          "reliable_ports": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reliable Ports"
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "DeploymentSummary"
      },
      "Diagnostic": {
        "properties": {
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "file": {
            "type": "string",
            "title": "File"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "severity",
          "file",
          "message"
        ],
        "title": "Diagnostic"
      },
      "DocumentFile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "content"
        ],
        "title": "DocumentFile"
      },
      "DocumentRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "title": "DocumentRef"
      },
      "DocumentUpdate": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "DocumentUpdate"
      },
      "DocumentWriteResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "bytes"
        ],
        "title": "DocumentWriteResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterfaceIndex": {
        "properties": {
          "interfaces": {
            "items": {
              "$ref": "#/components/schemas/InterfaceSummary"
            },
            "type": "array",
            "title": "Interfaces"
          },
          "deployments": {
            "items": {
              "$ref": "#/components/schemas/DeploymentSummary"
            },
            "type": "array",
            "title": "Deployments"
          }
        },
        "type": "object",
        "title": "InterfaceIndex"
      },
      "InterfaceSummary": {
        "properties": {
          "document": {
            "type": "string",
            "title": "Document"
          },
          "package": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "methods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Methods"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          },
          "structs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Structs"
          },
          "enums": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enums"
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "InterfaceSummary"
      },
...
      "NetworkSettings": {
        "properties": {
          "service_discovery": {
            "type": "boolean",
            "title": "Service Discovery",
            "default": true
          },
          "multicast": {
            "type": "string",
            "title": "Multicast",
            "default": "224.224.224.245"
          },
          "sd_port": {
            "type": "integer",
            "title": "Sd Port",
            "default": 30490
          }
        },
        "type": "object",
        "title": "NetworkSettings"
      },
...
      "ProjectCollection": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "type": "array",
            "title": "Projects"
          },
          "presets": {
            "items": {
              "$ref": "#/components/schemas/StarterPreset"
            },
            "type": "array",
            "title": "Presets"
          }
        },
        "type": "object",
        "title": "ProjectCollection"
      },
      "ProjectCreateRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "source_example": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Example"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "name"
        ],
        "title": "ProjectCreateRequest"
      },
...
      "ProjectOverview": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_example": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Example"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "readiness": {
            "type": "integer",
            "title": "Readiness"
          },
          "readiness_gates": {
            "items": {
              "$ref": "#/components/schemas/ReadinessGate"
            },
            "type": "array",
            "title": "Readiness Gates"
          },
          "recommended_action": {
            "$ref": "#/components/schemas/RecommendedAction"
          },
          "network": {
            "$ref": "#/components/schemas/NetworkSettings"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/TopologyNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "interface_index": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterfaceIndex"
              },
              {
                "type": "null"
              }
            ]
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentRef"
            },
            "type": "array",
            "title": "Documents"
          },
          "scenario_count": {
            "type": "integer",
            "title": "Scenario Count",
            "default": 0
          },
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Recent Runs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "readiness",
          "recommended_action",
          "network"
        ],
        "title": "ProjectOverview"
      },
      "ProjectStatus": {
        "properties": {
          "validation": {
            "type": "string",
            "title": "Validation"
          },
          "generation": {
            "type": "string",
            "title": "Generation"
          },
          "build": {
            "type": "string",
            "title": "Build"
          },
          "last_run": {
            "type": "string",
            "title": "Last Run"
          }
        },
        "type": "object",
        "required": [
          "validation",
          "generation",
          "build",
          "last_run"
        ],
        "title": "ProjectStatus"
      },
      "ProjectSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "default": 0
          },
          "node_count": {
            "type": "integer",
            "title": "Node Count",
            "default": 0
          },
          "scenario_count": {
            "type": "integer",
            "title": "Scenario Count",
            "default": 0
          },
          "latest_status": {
            "$ref": "#/components/schemas/ProjectStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "latest_status"
        ],
        "title": "ProjectSummary"
      },
      "ReadinessGate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "status",
          "ready",
          "detail"
        ],
        "title": "ReadinessGate"
      },
      "RecommendedAction": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "author",
              "build",
              "simulate",
              "inspect"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "title",
          "detail"
        ],
        "title": "RecommendedAction"
      },
...
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "capture": {
            "$ref": "#/components/schemas/CaptureSession"
          },
          "captures": {
            "items": {
              "$ref": "#/components/schemas/CaptureFile"
            },
            "type": "array",
            "title": "Captures"
          },
          "containers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ContainerState"
            },
            "type": "object",
            "title": "Containers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "scenario_id",
          "scenario_name",
          "status"
        ],
        "title": "RunSummary"
      },
...
      "StarterPreset": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_example": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Example"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "sample"
          },
          "runnable": {
            "type": "boolean",
            "title": "Runnable",
            "default": false
          },
          "default_project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Project Id"
          },
          "default_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "StarterPreset"
      },
...
      "TopologyNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "interface": {
            "type": "string",
            "title": "Interface"
          },
          "app_name": {
            "type": "string",
            "title": "App Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "interface",
          "app_name"
        ],
        "title": "TopologyNode"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationResult": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "interface_index": {
            "$ref": "#/components/schemas/InterfaceIndex"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "ValidationResult"
      },
...
    }
  }
}